Put off loading connection information
I had been loading connection information on create but then loading the data
when we get to the forEach loop. I now load the connection information when we
load the data. This fixes an issue where I declare a query object in an inline
and then try to use it outside the inline and get a "MySQL connection has gone
away" error.