No longer using a map but two staticarrays
The largest performance hit remaining was creating the map to be used by
trait_query_result, so I decided to keep the column names and the row data
seperate and lookup the data based on the corresponding position in the arrays.

Thus, instead of a map, trait_query_result now expects two
trait_positionallyKeyed (which it gets in the form of two staticarrays).

I have also defined the trati_positionallyKeyed->findFirst to find just= the
first element that matches the given value instead of using findPosition which
returns an array of all the positions that match. THIS IS A CHANGE IN
FUNCTIONALITY: Before, if I had multiple columns with the same name than a
trait_query_result would return an array of values, but now it returns the
value of the first matching column name.