Fixed a bug when looking up datasources that don't have a port
The code that looks up a datasource connection based on the hostname (and any
other specified parameters) make sure to cast the port as an integer. However,
if that meant the port got cast as 0, then we were passing '' for the port to
the other "onCreate" method, but that method requires an integer.